# openapi-live.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "s-curve",
    "description": "",
    "contact": {
      "name": "s-curve team"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8800"
    },
    {
      "url": "https://rust-service:8800"
    }
  ],
  "paths": {
...
    "/projects/{project_id}/progress": {
      "get": {
        "tags": [
          "Progress"
        ],
        "summary": "List Project Progress",
        "description": "List Project Progress Handles `GET` requests for `/projects/{project_id}/progress`. Requires bearer authentication.",
        "operationId": "list_project_progress",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "00000000-0000-0000-0000-000000000000"
          },
          {
            "name": "task_id",
            "in": "query",
            "description": "Optional task id filter.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            },
            "example": "00000000-0000-0000-0000-000000000000"
          }
        ],
        "responses": {
          "200": {
            "description": "List progress entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Progress"
                  }
                },
                "example": [
                  {
                    "created_at": "2025-01-19T09:00:00Z",
                    "id": "99999999-9999-4999-8999-999999999999",
                    "note": "Execution started",
                    "progress": 65,
                    "project_id": "44444444-4444-4444-8444-444444444444",
                    "task_id": "33333333-3333-4333-8333-333333333333",
                    "updated_at": "2025-01-19T09:00:00Z"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "description": "List Project Progress Handles `GET` requests for `/projects/{project_id}/progress`. Requires bearer authentication.",
        "operationId": "list_project_progress",
        "parameters": [
          {
            "description": "Project id",
            "example": "00000000-0000-0000-0000-000000000000",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Optional task id filter.",
            "example": "00000000-0000-0000-0000-000000000000",
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "created_at": "2025-01-19T09:00:00Z",
                    "id": "99999999-9999-4999-8999-999999999999",
                    "note": "Execution started",
                    "progress": 65,
                    "project_id": "44444444-4444-4444-8444-444444444444",
                    "task_id": "33333333-3333-4333-8333-333333333333",
                    "updated_at": "2025-01-19T09:00:00Z"
                  }
                ],
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Progress"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List progress entries"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Project Progress",
        "tags": [
          "Progress"
        ]
      }
    },
...
    "/projects/{project_id}/tasks/{task_id}/progress": {
      "get": {
        "tags": [
          "Progress"
        ],
        "summary": "List Progress",
        "description": "List Progress Handles `GET` requests for `/projects/{project_id}/tasks/{task_id}/progress`. Requires bearer authentication.",
        "operationId": "list_progress",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "00000000-0000-0000-0000-000000000000"
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "00000000-0000-0000-0000-000000000000"
          }
        ],
        "responses": {
          "200": {
            "description": "List progress entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Progress"
                  }
                },
                "example": [
                  {
                    "created_at": "2025-01-19T09:00:00Z",
                    "id": "99999999-9999-4999-8999-999999999999",
                    "note": "Execution started",
                    "progress": 65,
                    "project_id": "44444444-4444-4444-8444-444444444444",
                    "task_id": "33333333-3333-4333-8333-333333333333",
                    "updated_at": "2025-01-19T09:00:00Z"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Progress"
        ],
        "summary": "Create Progress",
        "description": "Create Progress Handles `POST` requests for `/projects/{project_id}/tasks/{task_id}/progress`. Requires bearer authentication.",
        "operationId": "create_progress",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "00000000-0000-0000-0000-000000000000"
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "00000000-0000-0000-0000-000000000000"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressCreateRequest"
              },
              "example": {
                "note": "Halfway there",
                "progress": 50
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Progress created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Progress"
                },
                "example": {
                  "created_at": "2025-01-19T09:00:00Z",
                  "id": "99999999-9999-4999-8999-999999999999",
                  "note": "Execution started",
                  "progress": 65,
                  "project_id": "44444444-4444-4444-8444-444444444444",
                  "task_id": "33333333-3333-4333-8333-333333333333",
                  "updated_at": "2025-01-19T09:00:00Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "description": "List Progress Handles `GET` requests for `/projects/{project_id}/tasks/{task_id}/progress`. Requires bearer authentication.",
        "operationId": "list_progress",
        "parameters": [
          {
            "description": "Project id",
            "example": "00000000-0000-0000-0000-000000000000",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Task id",
            "example": "00000000-0000-0000-0000-000000000000",
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "created_at": "2025-01-19T09:00:00Z",
                    "id": "99999999-9999-4999-8999-999999999999",
                    "note": "Execution started",
                    "progress": 65,
                    "project_id": "44444444-4444-4444-8444-444444444444",
                    "task_id": "33333333-3333-4333-8333-333333333333",
                    "updated_at": "2025-01-19T09:00:00Z"
                  }
                ],
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Progress"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List progress entries"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Progress",
        "tags": [
          "Progress"
        ]
      },
      "post": {
        "description": "Create Progress Handles `POST` requests for `/projects/{project_id}/tasks/{task_id}/progress`. Requires bearer authentication.",
        "operationId": "create_progress",
        "parameters": [
          {
            "description": "Project id",
            "example": "00000000-0000-0000-0000-000000000000",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Task id",
            "example": "00000000-0000-0000-0000-000000000000",
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "note": "Halfway there",
                "progress": 50
              },
              "schema": {
                "$ref": "#/components/schemas/ProgressCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2025-01-19T09:00:00Z",
                  "id": "99999999-9999-4999-8999-999999999999",
                  "note": "Execution started",
                  "progress": 65,
                  "project_id": "44444444-4444-4444-8444-444444444444",
                  "task_id": "33333333-3333-4333-8333-333333333333",
                  "updated_at": "2025-01-19T09:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Progress"
                }
              }
            },
            "description": "Progress created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Progress",
        "tags": [
          "Progress"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Progress": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "task_id",
          "progress",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "progress": {
            "type": "integer",
            "format": "int32"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "task_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProgressCreateRequest": {
        "type": "object",
        "required": [
          "progress"
        ],
        "properties": {
          "note": {
            "type": "string",
            "example": "Halfway done - waiting on review",
            "nullable": true
          },
          "progress": {
            "type": "integer",
            "format": "int32",
            "example": 75
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}